Implementing new method of storing credentials.  Added two utiltiy scripts and an invoicing script.
Database credentials will now be placed in creds.config.  This file at a minimum provides central management.  We should also be able to configure the web server never to serve this file (or place it in an unservable directory) for security.  Ideally we'd encrypt it too but one thing at a time.

Added two utility scripts.  clearLock.py will delete the row in the gnclock table which tells GnuCash the database is in use.  If a script transaction fails it often leavs the database locked and all further API commands will be rejected.  Running this wipes that row and lets you continue development.

sync-members reads the customers table and then creates rows in a new dh_members table which we will use to store custom member attributes.

invoice-multi-months allows the treasurer to create one or more invoices for a single member beginning with a future month.  Useful when a member pre-pays dues.